requirements-test.txt
# Testing dependencies for extraction pipeline
pytest>=7.4.0
pytest-benchmark>=4.0.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# Data validation and schema
jsonschema>=4.19.0
pydantic>=2.0.0

# Document processing
PyMuPDF>=1.23.0  # fitz for PDF text extraction
python-magic>=0.4.27  # MIME type detection
camelot-py[cv]>=0.10.1  # PDF table extraction
pdfplumber>=0.9.0  # PDF processing
openpyxl>=3.1.0  # Excel processing
pandas>=2.0.0  # Data manipulation

# OCR and image processing
pytesseract>=0.3.10  # OCR
Pillow>=10.0.0  # Image processing
opencv-python>=4.8.0  # Image preprocessing
ocrmypdf>=15.0.0  # PDF OCR

# AI and NLP
openai>=1.0.0  # OpenAI API
rapidfuzz>=3.0.0  # Fuzzy string matching
pint>=0.22  # Unit conversion

# Performance and monitoring
structlog>=23.0.0  # Structured logging
redis>=4.6.0  # Caching

# Utilities
tqdm>=4.65.0  # Progress bars
click>=8.1.0  # CLI interface